Moves GameBoard to 'model' package
