+++ app/src/main/java/com/example/bikesh/checkerz/GameState.java
@@ -1,5 +1,6 @@
 package com.example.bikesh.checkerz;
 
+import com.example.bikesh.checkerz.model.GameBoard;
 import com.example.bikesh.checkerz.model.PieceColor;
 
 import java.util.Collections;
