+++ app/src/main/java/com/example/bikesh/checkerz/model/GameBoard.java
@@ -1,5 +1,6 @@
-package com.example.bikesh.checkerz;
+package com.example.bikesh.checkerz.model;
 
+import com.example.bikesh.checkerz.Piece;
 import com.example.bikesh.checkerz.model.PieceColor;
 
 import java.util.ArrayList;
